tests: get show test working tests: move around some fixtures for access  feat: set config version in DeploymentMeta to default
