+++ pyproject.toml
@@ -81,7 +81,7 @@ exclude = [
 
 [tool.ruff.lint.per-file-ignores]
 '__init__.py' = ['E402']
-'tests/management/test_deployment_pointer.py' = ['E501']
+'tests/management/*' = ['E501']
 'tests/**/*' = [
     'T201',
     'S101',
